Adding SQL and GIF libraries to FIX >> Xcode IDE build (#2022)
* Adding SQL and GIF libraries to Xcode IDE build

* Updates to IGNORE

* Additional GIT help

* More GIT help

* Revert quilt pop changes

* Updates to help GIT

* Minor tweaks

* More build/git tweaks

* fixed warnings as errors (added casts)

* More tweaks

* Tweaks

* More ignore tweaks

* Finally ... Git Ignored External build artifacts

* Fixes

* Fixes for GIT and 'dukluv'

* Typo in path

* Fixed - one click builds  !!!!

* Changes to fix XCode IDE only build

* Warnings and Project changes

* Tweak

* Silence spurious error in 'external/build.sh` for Mac

* 1 more error in Xcode build script.

* One more warning.

* More warnings

+++ examples/pxScene2d/external/westeros-stub/westeros-stub.cpp
@@ -334,7 +334,7 @@ bool WstCompositorSetNestedDisplayName( WstCompositor *ctx, const char *nestedDi
 
       if ( nestedDisplayName )
       {
-         len= strlen(nestedDisplayName);
+         len= (int) strlen(nestedDisplayName);
 
          if ( (len == 0) || (len > MAX_NESTED_NAME_LEN) )
          {
@@ -935,7 +935,7 @@ bool WstCompositorLaunchClient( WstCompositor *ctx, const char *cmd )
          goto exit;
       }
 
-      int i= (cmd ? strlen(cmd) : 0);
+      int i= (cmd ? (int) strlen(cmd) : 0);
       if ( !cmd || (i > 255) )
       {
          sprintf( ctx->lastErrorDetail,
